requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
Authlib==1.3.0
httpx==0.27.0
itsdangerous==2.1.2
python-jose[cryptography]==3.3.0
APScheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.6.2
aiohttp==3.9.3
python-dateutil==2.9.0.post0
google-auth==2.28.1
requests==2.31.0
aiosqlite==0.20.0
watchdog==4.0.0
python-multipart==0.0.9
